% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{gd_enum_names}
\alias{gd_enum_names}
\alias{gd_enum_elements}
\alias{gd_resampling_methods}
\alias{gd_cloud_mask_methods}
\alias{gd_composite_methods}
\title{\code{geedim} Enums}
\usage{
gd_enum_names()

gd_enum_elements(enum = gd_enum_names())

gd_resampling_methods()

gd_cloud_mask_methods()

gd_composite_methods()
}
\arguments{
\item{enum}{Enum name, one or more of: \code{"CloudMaskMethod"}, \code{"CompositeMethod"}, \code{"ResamplingMethod"}}
}
\value{
\code{gd_enum_names()}: character vector containing names of Enums

\code{gd_enum_elements()}: element values of an Enum

\code{gd_resampling_methods()}: character vector of resampling methods (Enum \code{"ResamplingMethod"})

\code{gd_cloud_mask_methods()}: character vector of cloud mask methods (Enum \code{"CloudMaskMethod"})

\code{gd_composite_methods()}: character vector of composite methods (Enum \code{"CompositeMethod"})
}
\description{
\code{geedim} Enums
}
\examples{
\donttest{
if (gd_is_initialized())
  gd_enum_names()
}
\dontshow{if (gd_is_initialized()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
\donttest{
if (gd_is_initialized())
  gd_enum_elements()
}
\dontshow{if (gd_is_initialized()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
\donttest{
if (gd_is_initialized())
  gd_resampling_methods()
}
\dontshow{if (gd_is_initialized()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
\donttest{
if (gd_is_initialized())
  gd_cloud_mask_methods()
}
\dontshow{if (gd_is_initialized()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
\donttest{
if (gd_is_initialized())
  gd_composite_methods()
}
}
