% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{gd_get_asset}
\alias{gd_get_asset}
\alias{gd_update_asset}
\alias{gd_delete_asset}
\title{Get, Update, or Delete an Earth Engine Asset by ID}
\usage{
gd_get_asset(x)

gd_update_asset(x, asset, updateFields = character(0))

gd_delete_asset(x)
}
\arguments{
\item{x}{Asset ID name}

\item{asset}{Used only for \code{gd_update_asset()}, a list representing elements of \code{x} to replace. Only the \code{"start_time"}, \code{"end_time"}, and \code{"properties"} fields can be updated. If a value is named in "updateFields", but is unset in \code{asset}, then that value will be deleted from the asset.}

\item{updateFields}{Used only for \code{gd_update_asset()}. A character vector of field names to update. This may contain: "start_time" or "end_time" to update the corresponding timestamp, "properties.PROPERTY_NAME" to update a given property, or "properties" to update all properties. If the list is empty (default), all properties and both timestamps will be updated.}
}
\value{
This function is called for its side-effects (updates or deletes the specified asset)
}
\description{
Get, Update, or Delete an Earth Engine Asset by ID
}
\examples{
\dontrun{
a <- gd_get_asset("projects/your-project-name/assets/YOUR_ASSET_ID")
}
\dontrun{
a$properties$description <- "foo"
gd_update_asset("projects/your-project-name/assets/YOUR_ASSET_ID", a, "properties")
}
\dontrun{
gd_delete_asset("projects/your-project-name/assets/YOUR_ASSET_ID")
}
}
