% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_id.R
\name{gd_image_from_id}
\alias{gd_image_from_id}
\alias{gd_collection_from_id}
\title{Reference Google Earth Engine Image or ImageCollection by ID}
\usage{
gd_image_from_id(x)

gd_collection_from_id(x)
}
\arguments{
\item{x}{character. ID.}
}
\value{
an object referencing a Python \code{MaskedImage} or \code{MaskedCollection}, or \code{try-error} on error
}
\description{
Calls \code{geedim.mask.MaskedImage.from_id()} or \code{geedim.mask.MaskedCollection.from_id()}, respectively.
}
