% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gd}
\alias{gd}
\title{\code{Module(geedim)} - Get \code{geedim} Module Instance}
\usage{
gd()
}
\description{
Gets the \code{geedim} module instance in use by the package in current \strong{R}/\code{reticulate} session.
}
