% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gd_band_names}
\alias{gd_band_names}
\title{Get Names of Layers in an Earth Engine Image}
\usage{
gd_band_names(x)
}
\arguments{
\item{x}{a \code{geedim.image.ImageAccessor} (for geedim >= 2.0.0) or \code{geedim.download.BaseImage} (for geedim < 2.0.0) object, such as from \code{gd_image_from_id()}. See \verb{\link{geedim-versions}} for more details.}
}
\value{
character. Vector of names of each layer in an image.
}
\description{
Calls \code{bandNames()} method from \code{ee.Image} class.
}
\examples{
\dontshow{if (isTRUE(as.logical(Sys.getenv("R_RGEEDIM_RUN_EXAMPLES"))) && gd_is_initialized()) withAutoprint(\{ # examplesIf}
\donttest{
if (gd_is_initialized())
  gd_band_names(gd_image_from_id("USGS/SRTMGL1_003"))
}
\dontshow{\}) # examplesIf}
}
