% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gd_bandnames}
\alias{gd_bandnames}
\title{Get Names of Layers in an Earth Engine Image}
\usage{
gd_bandnames(x)
}
\arguments{
\item{x}{a Google Earth Engine Image object, such as from \code{gd_image_from_id()}}
}
\value{
character. Vector of names of each layer in an image.
}
\description{
Calls \code{bandNames()} method from \code{ee.Image} class.
}
