% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{gd_bbox}
\alias{gd_bbox}
\title{Prepare \code{region} Object for \code{download()}}
\usage{
gd_bbox(...)
}
\arguments{
\item{...}{either a single terra \code{SpatExtent} object or \emph{named} arguments \code{xmin}/\code{ymin}/\code{xmax}/\code{ymax}}
}
\value{
a \emph{list} object describing a GeoJSON bounding rectangular polygon suitable for use as \code{regions} argument to \code{download()}
}
\description{
Create a bounding box polygon Python object for use with \code{download()}. The coordinates of the bounding box are expressed in WGS84 decimal degrees (\code{"OGC:CRS84"}).
}
\examples{
gd_bbox(
  xmin = 5.744140,
  xmax = 6.528252,
  ymin = 49.44781,
  ymax = 50.18162
)
}
