% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gd_footprint}
\alias{gd_footprint}
\title{Get Footprint of Masked Image}
\usage{
gd_footprint(x)
}
\arguments{
\item{x}{a \code{geedim.mask.MaskedImage} object}
}
\value{
list.
}
\description{
Gets GeoJSON-style list containing footprint of a \code{geedim.mask.MaskedImage} object
}
