% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_clouds.R
\name{gd_mask_clouds}
\alias{gd_mask_clouds}
\title{Mask Clouds or Apply Fill Mask}
\usage{
gd_mask_clouds(x)
}
\arguments{
\item{x}{a \code{geedim.image.ImageAccessor} (for geedim >= 2.0.0) or \code{geedim.mask.MaskedImage} (for geedim < 2.0.0). See \verb{\link{geedim-versions}} for more details.}
}
\value{
a \code{geedim.image.ImageAccessor} (for geedim >= 2.0.0) or \code{geedim.mask.MaskedImage} (for geedim < 2.0.0). See \verb{\link{geedim-versions}} for more details.
}
\description{
Apply the cloud/shadow mask if supported, otherwise apply the fill mask.
}
