% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{get_projection}
\alias{get_projection}
\title{Get Projection Information from Google Earth Engine Asset}
\usage{
get_projection(x)
}
\arguments{
\item{x}{character ID referencing asset, or an image object (subclass of \code{ee.image.Image} or \code{geedim.download.BaseImage})}
}
\value{
\code{ee.Projection} object
}
\description{
Get Projection Information from Google Earth Engine Asset
}
