% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeedim-package.R
\docType{package}
\name{rgeedim-package}
\alias{rgeedim}
\alias{rgeedim-package}
\title{rgeedim: Search, Composite, and Download Google Earth Engine Imagery with the Python Module 'geedim'}
\description{
Search, composite, and download Google Earth Engine imagery with 'reticulate' bindings for the Python module 'geedim'. Wrapper functions are provided to make it more convenient to use 'geedim' to download images larger than the Earth Engine size limit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://humus.rocks/rgeedim/}
  \item \url{https://github.com/brownag/rgeedim}
  \item \url{https://geedim.readthedocs.io/}
  \item Report bugs at \url{https://github.com/brownag/rgeedim/issues}
}

}
\keyword{internal}
