% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\name{earthengine}
\alias{earthengine}
\alias{gd_ee_version}
\title{Get Earth Engine \code{Module(ee)} Instance}
\usage{
earthengine()

gd_ee_version()
}
\value{
character. Version Number.
}
\description{
Gets the \code{ee} module instance in use by \code{geedim} package in current session.

\code{gd_ee_version()} Gets the \code{ee} version using \code{importlib.metadata.version()}
}
