% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\name{earthengine}
\alias{earthengine}
\title{\code{Module(earthengine)} - Get \code{earthengine} Module Instance}
\usage{
earthengine()
}
\description{
Gets the \code{earthengine} module instance in use by the package in current \strong{R}/\code{reticulate} session.
}
