% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from.R
\name{gd_image_from_id}
\alias{gd_image_from_id}
\alias{gd_collection_from_name}
\alias{gd_collection_from_list}
\title{Reference Google Earth Engine Image or Image Collection by ID or Name}
\usage{
gd_image_from_id(x)

gd_collection_from_name(x)

gd_collection_from_list(x)
}
\arguments{
\item{x}{character. ID.}
}
\value{
reference to \code{geedim.MaskedImage} or \code{geedimMaskedCollection} object, or \code{try-error} on error
}
\description{
Create references to Images or Image Collections based on IDs and names, or combine image references into collections.
}
