% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gd_authenticate}
\alias{gd_authenticate}
\title{Authenticate with Google Earth Engine using \code{gcloud}}
\usage{
gd_authenticate(
  authorization_code = NULL,
  quiet = FALSE,
  code_verifier = NULL,
  auth_mode = NULL
)
}
\arguments{
\item{authorization_code}{Default: \code{NULL}}

\item{quiet}{Default: \code{FALSE}}

\item{code_verifier}{Default: \code{NULL}}

\item{auth_mode}{Default: \code{NULL}}
}
\description{
Calls \code{ee.Authenticate(...)}
}
