% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gd_band_names}
\alias{gd_band_names}
\title{Get Names of Layers in an Earth Engine Image}
\usage{
gd_band_names(x)
}
\arguments{
\item{x}{a Google Earth Engine Image object, such as from \code{gd_image_from_id()}}
}
\value{
character. Vector of names of each layer in an image.
}
\description{
Calls \code{bandNames()} method from \code{ee.Image} class.
}
\examples{
\dontshow{if (gd_is_initialized()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
if (gd_is_initialized())
  gd_band_names(gd_image_from_id("USGS/NED"))
}
