% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{gd_download}
\alias{gd_download}
\title{Download Google Earth Engine Image/ImageCollection}
\usage{
gd_download(
  x,
  filename = tempfile(fileext = ".tif"),
  region = NULL,
  overwrite = TRUE,
  silent = TRUE,
  ...
)
}
\arguments{
\item{x, }{ID, or a reference to \code{geedim.download.BaseImage}}

\item{filename}{path to file, defaults to temporary GeoTIFF file}

\item{region}{Default \code{NULL} Python object describing region}

\item{overwrite}{Overwrite existing file? Default: \code{TRUE}}

\item{silent}{Silence errors? Default: \code{TRUE}}

\item{...}{Additional arguments (e.g. \code{scale}) to \code{geedim.download.BaseImage$download(...)}}
}
\value{
(invisibly) path to downloaded image
}
\description{
Download Google Earth Engine Image/ImageCollection
}
