% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{gd_download}
\alias{gd_download}
\title{Download a Google Earth Engine Image}
\usage{
gd_download(
  x,
  filename = tempfile(fileext = ".tif"),
  region = NULL,
  overwrite = TRUE,
  silent = TRUE,
  ...
)
}
\arguments{
\item{x, }{ID or Name, or a reference to an object inheriting from \code{geedim.download.BaseImage} or \code{geedim.collection.MaskedCollection}}

\item{filename}{path to output file, defaults to temporary GeoTIFF file path}

\item{region}{a GeoJSON-like list, or other R spatial object describing region of interest, see \code{gd_region()} and \code{gd_bbox()} for details}

\item{overwrite}{Overwrite existing file? Default: \code{TRUE}}

\item{silent}{Silence errors? Default: \code{TRUE}}

\item{...}{Additional arguments (e.g. \code{scale}) to \code{geedim.download.BaseImage$download(...)}}
}
\value{
Invisible path to downloaded image, or \code{try-error} on error
}
\description{
Download a Google Earth Engine Image
}
\seealso{
\code{gd_region()} \code{gd_bbox()}
}
