% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gd_footprint}
\alias{gd_footprint}
\title{Get Footprint of Masked Image}
\usage{
gd_footprint(x)
}
\arguments{
\item{x}{a \code{geedim.image.ImageAccessor} (for geedim >= 2.0.0) or \code{geedim.mask.MaskedImage} (for geedim < 2.0.0) object. See \verb{\link{geedim-versions}} for more details.}
}
\value{
list.
}
\description{
Gets GeoJSON-style list containing footprint of a \code{geedim.image.ImageAccessor} (for geedim >= 2.0.0) or \code{geedim.mask.MaskedImage} (for geedim < 2.0.0) object. See \verb{\link{geedim-versions}} for more details.
}
\examples{
\dontshow{if (isTRUE(as.logical(Sys.getenv("R_RGEEDIM_RUN_EXAMPLES"))) && gd_is_initialized()) withAutoprint(\{ # examplesIf}
\donttest{
if (gd_is_initialized())
  gd_footprint(gd_image_from_id("USGS/SRTMGL1_003"))
}
\dontshow{\}) # examplesIf}
}
