% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gd_initialize}
\alias{gd_initialize}
\title{Initialize \code{geedim}
Calls \code{geedim} \code{Initialize()} method}
\usage{
gd_initialize(quiet = TRUE)
}
\arguments{
\item{quiet}{Suppress error messages on load? Default: \code{TRUE}}
}
\value{
try-error (invisibly) on error
}
\description{
Initialize \code{geedim}
Calls \code{geedim} \code{Initialize()} method
}
\examples{
gd_initialize()
}
