% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gd_initialize}
\alias{gd_initialize}
\alias{gd_is_initialized}
\title{Initialize \code{geedim}}
\usage{
gd_initialize(
  private_key_file = NULL,
  opt_url = "https://earthengine-highvolume.googleapis.com",
  quiet = TRUE
)

gd_is_initialized()
}
\arguments{
\item{private_key_file}{Optional: Path to JSON file containing client information and private key.}

\item{opt_url}{Base URL for API requests; defaults to "High Volume": \code{"https://earthengine-highvolume.googleapis.com"}}

\item{quiet}{Suppress error messages on load? Default: \code{FALSE}}
}
\value{
\code{gd_initialize()}: try-error (invisibly) on error.

\code{gd_is_initialized()}: logical. \code{TRUE} if initialized successfully.
}
\description{
Calls \code{geedim} \code{Initialize()} method. This method should be called at the beginning of each session.
}
\examples{
\donttest{
gd_initialize()
}
gd_is_initialized()
}
\seealso{
\code{gd_authenticate()}
}
