% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_clouds.R
\name{gd_mask_clouds}
\alias{gd_mask_clouds}
\title{Mask Clouds or Apply Fill Mask}
\usage{
gd_mask_clouds(x)
}
\arguments{
\item{x}{a \code{geedim.mask.MaskedImage}}
}
\value{
a \code{geedim.mask.MaskedImage}
}
\description{
Apply the cloud/shadow mask if supported, otherwise apply the fill mask.
}
