% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{gd_projection}
\alias{gd_projection}
\title{Get Projection Information from Google Earth Engine Asset}
\usage{
gd_projection(x)
}
\arguments{
\item{x}{character ID referencing asset, or an image object (subclass of \code{ee.image.Image}, \code{geedim.image.ImageAccessor} for geedim >= 2.0.0, or \code{geedim.download.BaseImage} for geedim < 2.0.0). See \verb{\link{geedim-versions}} for more details.}
}
\value{
\code{ee.Projection} object
}
\description{
Get Projection Information from Google Earth Engine Asset
}
\examples{
\dontshow{if (isTRUE(as.logical(Sys.getenv("R_RGEEDIM_RUN_EXAMPLES"))) && gd_is_initialized()) withAutoprint(\{ # examplesIf}
\donttest{
if (gd_is_initialized())
  gd_projection(gd_image_from_id('CSP/ERGo/1_0/Global/SRTM_topoDiversity'))
}
\dontshow{\}) # examplesIf}
}
