% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{gd_projection}
\alias{gd_projection}
\title{Get Projection Information from Google Earth Engine Asset}
\usage{
gd_projection(x)
}
\arguments{
\item{x}{character ID referencing asset, or an image object (subclass of \code{ee.image.Image} or \code{geedim.download.BaseImage})}
}
\value{
\code{ee.Projection} object
}
\description{
Get Projection Information from Google Earth Engine Asset
}
\examples{
\dontshow{if (length(geedim()) > 1) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\donttest{
gd_initialize()
gd_projection(gd_image_from_id('CSP/ERGo/1_0/Global/SRTM_topoDiversity'))
}
\dontshow{\}) # examplesIf}
}
