% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{gd_properties}
\alias{gd_properties}
\title{Get Properties of an Image Collection}
\usage{
gd_properties(x)
}
\arguments{
\item{x}{\code{geedim.collection.MaskedCollection} object}
}
\value{
\code{data.frame} containing properties table from \code{x}; \code{NULL} if no properties table.
}
\description{
Get Properties of an Image Collection
}
