% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{gd_region}
\alias{gd_region}
\title{Create \code{region} Object from R Spatial Objects}
\usage{
gd_region(x)
}
\arguments{
\item{x}{either a WKT string (character), a {terra} SpatRaster/SpatVector, an {sf} object, a {sp} Spatial* object or a {raster} RasterLayer/Stack.}
}
\value{
result methods as
}
\description{
Creates a suitable input for the \code{region} argument to \verb{gd_download(<Image>)} or \code{gd_search()} for Image Collections.
}
\seealso{
\code{gd_bbox()}
}
